+++ benches/fibonacci.rs
@@ -7,7 +7,7 @@ use halo2curves::bn256::Fr as Bn;
 use std::{sync::Arc, time::Duration};
 
 use lurk::{
-    eval::lang::{Coproc, Lang},
+    lang::{Coproc, Lang},
     lem::{eval::evaluate, store::Store},
     proof::nova::NovaProver,
     public_parameters::{
