+++ benches/sha256.rs
@@ -15,8 +15,8 @@ use std::{sync::Arc, time::Duration};
 
 use lurk::{
     coprocessor::sha256::{Sha256Coproc, Sha256Coprocessor},
-    eval::lang::Lang,
     field::LurkField,
+    lang::Lang,
     lem::{
         eval::{evaluate, make_cprocs_funcs_from_lang, make_eval_step_from_config, EvalConfig},
         pointers::Ptr,
