+++ src/lem/tests/nivc_steps.rs
@@ -2,7 +2,7 @@ use halo2curves::bn256::Fr;
 
 use crate::{
     coprocessor::test::DumbCoprocessor,
-    eval::lang::Lang,
+    lang::Lang,
     lem::{
         eval::{evaluate, make_cprocs_funcs_from_lang, make_eval_step_from_config, EvalConfig},
         store::{expect_ptrs, intern_ptrs, Store},
