+++ src/lib.rs
@@ -8,9 +8,9 @@ pub mod config;
 pub mod coprocessor;
 pub mod coroutine;
 pub mod error;
-pub mod eval;
 pub mod field;
 mod hash;
+pub mod lang;
 pub mod lem;
 mod num;
 mod package;
